Refine Moddy script
